pyproject.toml
[project]
name = "recipe-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.116.1",
    "minsearch>=0.0.4",
    "numpy>=2.2.6",
    "openai>=1.99.9",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "qdrant-client[fastembed]>=1.14.2",
    "requests>=2.32.4",
]
